@@ -4,28 +4,26 @@
   "private": true,
   "homepage": "https://studentvlad5.github.io/finance-club/",
   "dependencies": {
-    "@reduxjs/toolkit": "^1.9.5",
-    "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^14.0.0",
+    "@reduxjs/toolkit": "^1.9.7",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^14.1.0",
     "@testing-library/user-event": "^13.5.0",
     "aos": "^2.3.4",
-    "axios": "^1.4.0",
-    "formik": "^2.4.0",
+    "axios": "^1.6.1",
+    "formik": "^2.4.5",
     "modern-normalize": "^2.0.0",
     "notiflix": "^3.2.6",
     "rc-pagination": "^3.7.0",
-    "react": "^18.0.2",
+    "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-helmet-async": "^1.3.0",
-    "react-hot-toast": "^2.4.1",
     "react-icons": "^4.11.0",
-    "react-redux": "^8.0.5",
-    "react-router-dom": "^6.11.1",
+    "react-redux": "^8.1.3",
+    "react-router-dom": "^6.18.0",
     "react-scripts": "^5.0.1",
     "redux": "^4.2.1",
     "redux-persist": "^6.0.0",
-    "styled-components": "^6.0.0-rc.1",
-    "typescript": "^3.2.1",
+    "styled-components": "^6.1.1",
     "web-vitals": "^2.1.3",
     "yup": "^1.3.2"
   },
@@ -68,7 +66,7 @@
     "club",
     "project"
   ],
-  "author": "Vlad Popov",
+  "author": "Brand maze",
   "license": "UNLICENSED",
   "bugs": {
     "url": "https://github.com/StudentVlad5/finance-club/issues"
