@@ -1,4 +1,15 @@
-import styled from 'styled-components';
+import styled, { keyframes } from 'styled-components';
+
+const fadeInTopAnimation = keyframes`
+  0% {
+    transform: translateY(-50px);
+    opacity: 0;
+  }
+  100% {
+    transform: translateY(0);
+    opacity: 1;
+  }
+`;
 
 const BtnChangeTheme = styled.button`
   background-color: transparent;
@@ -21,6 +32,7 @@ const SwitcherWrapper = styled.div`
   height: auto;
 
   background-color: transparent;
+  animation: ${fadeInTopAnimation} 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) both;
   cursor: pointer;
 
   @media screen and (min-width: 768px) {
