@@ -1,9 +1,7 @@
-import axios from "axios";
+import axios from 'axios';
+import { BASE_URL } from 'helpers/constants';
 
-const BASE_URL = "https://rich-rose-shoulder-pads.cyclic.app/api";
-// const BASE_URL = "http://localhost:3030/api";
-
-export const signUp = async (credentials) => {
+export const signUp = async credentials => {
   try {
     const res = await axios.post(`${BASE_URL}/auth/signup`, credentials);
     return res;
@@ -12,7 +10,7 @@ export const signUp = async (credentials) => {
   }
 };
 
-export const signIn = async (credentials) => {
+export const signIn = async credentials => {
   try {
     const res = await axios.post(`${BASE_URL}/auth/signin`, credentials);
     return res;
@@ -26,21 +24,21 @@ export const singOut = async () => {
   return res;
 };
 
-export const updateUserData = async (updateData) => {
+export const updateUserData = async updateData => {
   const asArray = Object.entries(updateData);
-  const filtered = asArray.filter(([key]) => key !== "_id");
+  const filtered = asArray.filter(([key]) => key !== '_id');
   const justOne = Object.fromEntries(filtered);
   const { data } = await axios.patch(
     `${BASE_URL}/auth/user/${updateData._id}`,
     justOne,
     {
       headers: {
-        "Content-Type": "multipart/form-data",
-        "Access-Control-Allow-Origin": "*",
-        "Access-Control-Allow-Methods": "GET,POST,PUT,PATCH,DELETE,OPTIONS",
-        "Access-Control-Expose-Headers": "Content-Range",
+        'Content-Type': 'multipart/form-data',
+        'Access-Control-Allow-Origin': '*',
+        'Access-Control-Allow-Methods': 'GET,POST,PUT,PATCH,DELETE,OPTIONS',
+        'Access-Control-Expose-Headers': 'Content-Range',
       },
-    }
+    },
   );
   return data;
 };
@@ -48,4 +46,40 @@ export const updateUserData = async (updateData) => {
 export const refreshUserToken = async () => {
   const { data } = await axios.post(`${BASE_URL}/auth`);
   return data;
+};
+
+export const changePassword = async credentials => {
+  try {
+    const res = await axios.post('/auth/changePassword', credentials);
+    return res;
+  } catch (error) {
+    return error.message;
+  }
+};
+
+export const forgotPassword = async credentials => {
+  try {
+    const res = await axios.post('/auth/forgotPassword', credentials);
+    return res;
+  } catch (error) {
+    return error.message;
+  }
+};
+
+export const addToEvents = async id => {
+  try {
+    await axios.post(`/events/${id}`);
+    return id;
+  } catch (error) {
+    return error.message;
+  }
+};
+
+export const removeFromEvents = async id => {
+  try {
+    await axios.delete(`/events/${id}`);
+    return id;
+  } catch (error) {
+    return error.message;
+  }
 };
