@@ -1,22 +1,18 @@
-import axios from "axios";
-import { refreshUser } from "../redux/auth/operations";
-import { store } from "../redux/store";
-
-// const { BASE_URL } = window.global;
-
-const BASE_URL = "https://rich-rose-shoulder-pads.cyclic.app/api";
-// const BASE_URL = "http://localhost:3030/api";
+import axios from 'axios';
+import { refreshUser } from '../redux/auth/operations';
+import { store } from '../redux/store';
+import { BASE_URL } from 'helpers/constants';
 
 axios.defaults.baseURL = BASE_URL;
-export * from "./auth";
+export * from './auth';
 
 axios.interceptors.request.use(
   function (config) {
     return config;
   },
   function (error) {
-    return Promise.reject("Error: ", error);
-  }
+    return Promise.reject('Error: ', error);
+  },
 );
 
 axios.interceptors.request.use(
@@ -27,6 +23,6 @@ axios.interceptors.request.use(
     if (error.response.status === 401) {
       store.dispatch(refreshUser());
     }
-    return Promise.reject("Error: ", error);
-  }
+    return Promise.reject('Error: ', error);
+  },
 );
