@@ -1,13 +1,12 @@
-import axios from "axios";
-const BASE_URL = "https://rich-rose-shoulder-pads.cyclic.app/api";
-// const BASE_URL = "http://localhost:3030/api";
+import axios from 'axios';
+import { BASE_URL } from 'helpers/constants';
 
 export function leaveMessage(body) {
   return axios.post(`${BASE_URL}/message`, body, {
     headers: {
-      "content-type": "application/json",
-      "Access-Control-Allow-Origin": "*",
-      "Access-Control-Allow-Methods": "GET,POST,PUT,PATCH,DELETE,OPTIONS",
+      'content-type': 'application/json',
+      'Access-Control-Allow-Origin': '*',
+      'Access-Control-Allow-Methods': 'GET,POST,PUT,PATCH,DELETE,OPTIONS',
     },
   });
 }
